requirements.txt
alembic==1.4.3
attrs==20.2.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
